post서비스 체크

+++ src/main/java/com/example/projectpicker/post/dto/response/PageResponseDTO.java
@@ -32,7 +32,7 @@ public class PageResponseDTO<T> {
     public PageResponseDTO(Page<T> pageData){
         this.totalCount = (int) pageData.getTotalElements();
         this.currentPage = pageData.getPageable().getPageNumber() + 1;
-        this.endPage = (int) (Math.ceil((double) currentPage) / (PAGE_COUNT) * PAGE_COUNT);
+        this.endPage = (int) (Math.ceil(currentPage) / (PAGE_COUNT) * PAGE_COUNT);
         this.startPage = endPage - PAGE_COUNT + 1;
 
         // 페이지 마지막 구간에 endPage 값 보정
